Started epic refactoring: - contents of structs and arrays are now symbolic heaps - operations on structs and arrays now unified with heapy ones - regrouped utils, code splitted into Prelude and Collections modules
