+++ VSharp.SILI/State.fs
@@ -2,91 +2,74 @@
 open VSharp.Utils
 
 module internal State =
-    [<StructuralEquality;CustomComparison>]
-    type internal HeapKey =
-        | ConcreteAddress of int
-        | StaticAddress of string
-        | SymbolicAddress of string * SymbolicConstantSource
-        with
-        interface System.IComparable with
-            member x.CompareTo y =
-                match y with
-                | :? HeapKey as k ->
-                    match x, k with
-                    | ConcreteAddress a1, ConcreteAddress a2 -> compare a1 a2
-                    | ConcreteAddress _, _ -> -1
-                    | StaticAddress _, ConcreteAddress _ -> 1
-                    | StaticAddress a1, StaticAddress a2 -> compare a1 a2
-                    | StaticAddress _, _ -> -1
-                    | SymbolicAddress(a1, _), SymbolicAddress(a2, _) -> compare a1 a2
-                    | SymbolicAddress _, _ -> 1
-                | t -> internalfail (sprintf "comparing heap key with %s" (toString t))
-        override this.ToString() =
-            match this with
-            | ConcreteAddress a -> a.ToString()
-            | StaticAddress a -> System.Type.GetType(a).FullName
-            | SymbolicAddress(a, _) -> a
-
-    type internal StackKey = string * string // name and token
+    module SymbolicHeap = Heap
 
     type internal stack = MappedStack.stack<StackKey, Term>
-    type internal heap = Map<HeapKey, Term>
+    type internal heap = SymbolicHeap
+    type internal staticMemory = SymbolicHeap
     type internal frames = Stack.stack<StackKey list>
-    type internal path = Term list
-    type internal state = stack * heap * frames * path
+    type internal pathCondition = Term list
+    type internal state = stack * heap * staticMemory * frames * pathCondition
 
-    let internal empty : state = (MappedStack.empty, Map.empty, Stack.empty, List.empty)
+    let internal empty : state = (MappedStack.empty, SymbolicHeap.empty, SymbolicHeap.empty, Stack.empty, List.empty)
 
-    let internal derefStack (s, _, _, _) key = MappedStack.find key s
-    let internal pushToStack (s, _, _, _) key value = MappedStack.push key value s
-    let internal mutateStack (s, _, _, _) key value = MappedStack.add key value s
-    let internal isAllocatedOnStack (s, _, _, _) key = MappedStack.containsKey key s
-    let internal stackMap f (s, _, _, _) = MappedStack.map f s
-    let internal compareStacks = MappedStack.compare
-
-    let internal push ((s, h, f, p) : state) frame : state =
+    let internal readStackLocation ((s, _, _, _, _) : state) key = MappedStack.find key s
+    let internal newStackFrame ((s, h, m, f, p) : state) frame : state =
         let pushOne (map : stack) (key, value) = (key, MappedStack.push key value map)
         let names, newStack = frame |> List.mapFold pushOne s in
-        (newStack, h, Stack.push f names, p)
-    let internal pop ((s, h, f, p) : state) : state =
+        (newStack, h, m, Stack.push f names, p)
+    let internal popStack ((s, h, m, f, p) : state) : state =
         let popOne (map : stack) name = MappedStack.remove map name
         let names = Stack.peak f in
-        (List.fold popOne s names, h, Stack.pop f, p)
+        (List.fold popOne s names, h, m, Stack.pop f, p)
+    let internal pushToCurrentStackFrame ((s, _, _, _, _) : state) key value = MappedStack.push key value s
+    let internal mutateStack ((s, _, _, _, _) : state) key value = MappedStack.add key value s
+    let internal isAllocatedOnStack ((s, _, _, _, _) : state) key = MappedStack.containsKey key s
+    let internal compareStacks s1 s2 = MappedStack.compare (fun key value -> StackRef(key, [], PointerType (Terms.TypeOf value))) s1 s2
 
-    let internal stack ((s, _, _, _) : state) = s
-    let internal pathCondition ((_, _, _, p) : state) = p
-    let internal frames ((_, _, f, _) : state) = f
+    let internal readHeapLocation ((_, h, _, _, _) : state) key = h.[key]
+    let internal readStaticLocation ((_, _, m, _, _) : state) key = m.[key]
+    let internal staticMembersInitialized ((_, _, m, _, _) : state) typeName =
+        SymbolicHeap.contains (Concrete(typeName, String)) m
 
-    let internal withPathCondition ((s, h, f, p) : state) cond : state = (s, h, f, cond::p)
-    let internal popPathCondition ((s, h, f, p) : state) : state =
+    let internal withPathCondition ((s, h, m, f, p) : state) cond : state = (s, h, m, f, cond::p)
+    let internal popPathCondition ((s, h, m, f, p) : state) : state =
         match p with
         | [] -> internalfail "cannot pop empty path condition"
-        | _::p' -> (s, h, f, p')
+        | _::p' -> (s, h, m, f, p')
 
-    let private mergeMaps map1 map2 find add fold contains resolve =
-        let resolveIfShould map key value =
-            if contains key map then
-                let oldValue = find key map in
-                let newValue = value in
-                if oldValue = newValue then map
-                else
-                    add key (resolve oldValue newValue) map
-            else
-                add key value map
-        fold resolveIfShould map1 map2
+    let private stackOf ((s, _, _, _, _) : state) = s
+    let private heapOf ((_, h, _, _, _) : state) = h
+    let private staticsOf ((_, _, m, _, _) : state) = m
+    let private framesOf ((_, _, _, f, _) : state) = f
+    let internal pathConditionOf ((_, _, _, _, p) : state) = p
 
-    let internal merge ((s1, h1, f1, p1) : state) ((s2, h2, f2, p2) : state) resolve : state =
+    let internal merge ((s1, h1, m1, f1, p1) : state) ((s2, h2, m2, f2, p2) : state) resolve : state =
         assert(p1 = p2)
         assert(f1 = f2)
-        let mergedStack = mergeMaps s1 s2 MappedStack.find MappedStack.add MappedStack.fold MappedStack.containsKey resolve in
-        let mergedHeap = mergeMaps h1 h2 Map.find Map.add Map.fold Map.containsKey resolve in
-        (mergedStack, mergedHeap, f1, p1)
+        let mergedStack = MappedStack.merge2 s1 s2 resolve in
+        let mergedHeap = Heap.merge2 h1 h2 resolve in
+        let mergedStatics = Heap.merge2 m1 m2 resolve in
+        (mergedStack, mergedHeap, mergedStatics, f1, p1)
 
-    let internal staticMembersInitialized ((_, h, _, _) : state) typeName =
-        h.ContainsKey(StaticAddress typeName)
+    let internal mergeMany guards states resolve : state =
+        assert(List.length states > 0)
+        let first = List.head states in
+        let frames = framesOf first in
+        let path = pathConditionOf first in
+        assert(List.forall (fun s -> framesOf s = frames) states)
+        assert(List.forall (fun s -> pathConditionOf s = path) states)
+        let mergedStack = MappedStack.merge guards (List.map stackOf states) resolve in
+        let mergedHeap = Heap.merge guards (List.map heapOf states) resolve in
+        let mergedStatics = Heap.merge guards (List.map staticsOf states) resolve in
+        (mergedStack, mergedHeap, mergedStatics, frames, path)
 
-    let internal dumpHeap ((_, h, _, _) : state) =
-        let elements = h |> Map.toList |> List.map (fun (key, value) ->
-            key.ToString() + " ==> " + value.ToString())
-        in
-        List.sort elements |> join "\n"
+    let private staticKeyToString = function
+        | Concrete(typeName, String) -> System.Type.GetType(typeName :?> string).FullName
+        | t -> toString t
+
+    let internal dumpMemory ((_, h, m, _, _) : state) =
+        let sh = Heap.dump h toString in
+        let mh = Heap.dump m staticKeyToString in
+        let separator = if System.String.IsNullOrWhiteSpace(sh) then "" else "\n"
+        sh + separator + mh
