+++ VSharp.Utils/VSharp.Utils.fsproj
@@ -35,6 +35,9 @@
     <DocumentationFile>bin\Release\VSharp.Utils.XML</DocumentationFile>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="FSharpx.Collections, Version=1.15.2.0, Culture=neutral, PublicKeyToken=null">
+      <HintPath>..\packages\FSharpx.Collections.1.15.2\lib\net40\FSharpx.Collections.dll</HintPath>
+    </Reference>
     <Reference Include="mscorlib" />
     <Reference Include="FSharp.Core, Version=$(TargetFSharpCoreVersion), Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a">
       <Private>True</Private>
@@ -45,11 +48,11 @@
   <ItemGroup>
     <Compile Include="AssemblyInfo.fs" />
     <Compile Include="IdGenerator.fs" />
-    <Compile Include="ActivePatterns.fs" />
-    <Compile Include="Wrappers.fs" />
-    <Compile Include="MappedStack.fs" />
+    <Compile Include="Prelude.fs" />
+    <Compile Include="Collections.fs" />
     <Compile Include="Cps.fs" />
-    <Compile Include="Stack.fs" />
+    <Compile Include="MappedStack.fs" />
+    <Compile Include="Heap.fs" />
   </ItemGroup>
   <PropertyGroup>
     <MinimumVisualStudioVersion Condition="'$(MinimumVisualStudioVersion)' == ''">11</MinimumVisualStudioVersion>
